Refactor:  더미데이터 ts 파일화 및 index.ts 경로 정리
