+++ src/components/mainpage/CoursesData.ts
