+++ src/components/mainpage/index.ts
@@ -1,0 +1,2 @@
+export { FeaturesData } from './FeaturesData'
+export { CoursesData } from './CoursesData'
